chore: refactor type exports for improved modularity in interface files
