+++ src/business/utils/parsed-image-reference.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export interface ParsedImageReference {
+  registry: string;
+  repository: string;
+  tag: string;
+}
