+++ src/commands/network-deploy-config-class.ts
@@ -1,0 +1,76 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type NamespaceName} from '../types/namespace/namespace-name.js';
+import {type NodeAlias, type NodeAliases, type IP} from '../types/aliases.js';
+import {type ClusterReferenceName, type ClusterReferences} from '../types/index.js';
+import {type HelmChartValues} from '../integration/helm/model/values.js';
+import {type StorageType} from '../core/constants.js';
+import {type ConsensusNode} from '../core/model/consensus-node.js';
+import {type BlockNodeStateSchema} from '../data/schema/model/remote/state/block-node-state-schema.js';
+
+export interface NetworkDeployConfigClass {
+  isUpgrade: boolean;
+  applicationEnv: string;
+  chainId: string;
+  cacheDir: string;
+  chartDirectory: string;
+  loadBalancerEnabled: boolean;
+  soloChartVersion: string;
+  namespace: NamespaceName;
+  deployment: string;
+  nodeAliasesUnparsed: string;
+  persistentVolumeClaims: string;
+  releaseTag: string;
+  keysDir: string;
+  nodeAliases: NodeAliases;
+  stagingDir: string;
+  stagingKeysDir: string;
+  valuesFile: string;
+  chartValuesMap: Record<ClusterReferenceName, HelmChartValues>;
+  grpcTlsCertificatePath: string;
+  grpcWebTlsCertificatePath: string;
+  grpcTlsKeyPath: string;
+  grpcWebTlsKeyPath: string;
+  genesisThrottlesFile: string;
+  resolvedThrottlesFile: string;
+  haproxyIps: string;
+  envoyIps: string;
+  haproxyIpsParsed?: Record<NodeAlias, IP>;
+  envoyIpsParsed?: Record<NodeAlias, IP>;
+  storageType: StorageType;
+  gcsWriteAccessKey: string;
+  gcsWriteSecrets: string;
+  gcsEndpoint: string;
+  gcsBucket: string;
+  gcsBucketPrefix: string;
+  awsWriteAccessKey: string;
+  awsWriteSecrets: string;
+  awsEndpoint: string;
+  awsBucket: string;
+  awsBucketPrefix: string;
+  awsBucketRegion: string;
+  backupBucket: string;
+  backupWriteSecrets: string;
+  backupWriteAccessKey: string;
+  backupEndpoint: string;
+  backupRegion: string;
+  backupProvider: string;
+  consensusNodes: ConsensusNode[];
+  contexts: string[];
+  clusterRefs: ClusterReferences;
+  domainNames?: string;
+  domainNamesMapping?: Record<NodeAlias, string>;
+  blockNodeComponents: BlockNodeStateSchema[];
+  debugNodeAlias: NodeAlias;
+  app: string;
+  serviceMonitor: string;
+  podLog: string;
+  singleUseServiceMonitor: string;
+  singleUsePodLog: string;
+  enableMonitoringSupport: boolean;
+  javaFlightRecorderConfiguration: string;
+  wrapsEnabled: boolean;
+  wrapsKeyPath: string;
+  tssEnabled: boolean;
+  minioEnabled: boolean;
+}
