+++ src/commands/network-destroy-context.ts
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type NamespaceName} from '../types/namespace/namespace-name.js';
+
+export interface NetworkDestroyContext {
+  config: {
+    deletePvcs: boolean;
+    deleteSecrets: boolean;
+    namespace: NamespaceName;
+    enableTimeout: boolean;
+    force: boolean;
+    contexts: string[];
+    deployment: string;
+  };
+  checkTimeout: boolean;
+}
