+++ src/commands/node/config-interfaces/checked-nodes-context.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type CheckedNodesConfigClass} from './checked-nodes-config-class.js';
+
+export interface CheckedNodesContext {
+  config: CheckedNodesConfigClass;
+}
