+++ src/commands/node/config-interfaces/node-common-config-class.ts
@@ -1,10 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {type NamespaceName} from '../../../types/namespace/namespace-name.js';
-import {type NodeAlias, type NodeAliases} from '../../../types/aliases.js';
 import {type ConsensusNode} from '../../../core/model/consensus-node.js';
-import {type PodReference} from '../../../integration/kube/resources/pod/pod-reference.js';
-import {type NodeServiceMapping} from '../../../types/mappings/node-service-mapping.js';
 
 export interface NodeCommonConfigClass {
   namespace: NamespaceName;
@@ -16,23 +13,7 @@ export interface NodeCommonConfigClass {
   stagingDir: string;
 }
 
-export interface NodeCommonConfigWithNodeAliases extends NodeCommonConfigClass {
-  nodeAliases: NodeAliases;
-  nodeAliasesUnparsed: string;
-}
-
-export interface NodeCommonConfigWithNodeAlias extends NodeCommonConfigClass {
-  nodeAlias: NodeAlias;
-}
-
-export interface CheckedNodesConfigClass extends NodeCommonConfigClass {
-  podRefs: Record<NodeAlias, PodReference>;
-  skipStop: boolean;
-  existingNodeAliases: NodeAliases;
-  allNodeAliases: NodeAliases;
-  serviceMap: NodeServiceMapping;
-}
-
-export interface CheckedNodesContext {
-  config: CheckedNodesConfigClass;
-}
+export {type NodeCommonConfigWithNodeAliases} from './node-common-config-with-node-aliases.js';
+export {type NodeCommonConfigWithNodeAlias} from './node-common-config-with-node-alias.js';
+export {type CheckedNodesConfigClass} from './checked-nodes-config-class.js';
+export {type CheckedNodesContext} from './checked-nodes-context.js';
