+++ src/commands/node/config-interfaces/node-common-config-with-node-alias.ts
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type NodeAlias} from '../../../types/aliases.js';
+import {type NodeCommonConfigClass} from './node-common-config-class.js';
+
+export interface NodeCommonConfigWithNodeAlias extends NodeCommonConfigClass {
+  nodeAlias: NodeAlias;
+}
