+++ src/commands/node/config-interfaces/node-common-config-with-node-aliases.ts
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type NodeAliases} from '../../../types/aliases.js';
+import {type NodeCommonConfigClass} from './node-common-config-class.js';
+
+export interface NodeCommonConfigWithNodeAliases extends NodeCommonConfigClass {
+  nodeAliases: NodeAliases;
+  nodeAliasesUnparsed: string;
+}
